backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
asyncpg==0.30.0
python-dotenv==1.0.0
openai==1.3.0
httpx==0.25.2
websockets==12.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0
email-validator>=2.0